\name{DeutscheBank}
\alias{DeutscheBank}
\docType{data}
\title{Daily final prices (DAX) of the German stock Deutsche Bank in the years 2006 and 2007}
\description{Containing the daily final prices of the German stock Deutsche Bank in the years 2006 and 2007.}

\usage{data(DeutscheBank)}

\format{
A data frame with 507 observations of the following 2 variables.
  \describe{
	\item{\code{Date}}{Date}
	\item{\code{ClosingPrice}}{ClosingPrice}
  }
}

\keyword{datasets}