\name{pendenForm}
\alias{pendenForm}
\title{Formula interpretation and data transfer}
\description{
Function 'pendenForm' interprets the input 'form' of the function pencopula(),transfers the data back to the main program.
}
\usage{
pendenForm(penden.env)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{penden.env}{environment used in pendensity()}
}

\value{
Returning the values 'Y', the number of values 'n' and covariates 'p'.
}
\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}
\references{Flexible Copula Density Estimation with Penalized
  Hierarchical B-Splines, Kauermann G., Schellhase C. and Ruppert, D. (2013), Scandinavian Journal of Statistics 40(4), 685-705.}
\keyword{IO}
